Makes a protocol that describes a building, adding various properties and methods, then create two structs, House and Office. The protocol includes the following:
A property storing how many rooms it has.
A property storing the cost as an integer (e.g. 500,000 for a building costing $500,000.)
A property storing the name of the estate agent responsible for selling the building.
A method for printing the sales summary of the building, describing what it is along with its other properties.
An extension that adds an addtional room if needed.